update order controller with apperror

+++ src/controllers/item.js
@@ -57,7 +57,12 @@ const deleteItem = async(req, res) => {
 
     try {
         const { id } = req.params;
-        const item = await Item.findByIdAndDelete(id);
+        const item = await Item.findById(id);
+        if (!item) {
+            return res.status(404)
+                      .json({ message: 'Item not found' });
+        }
+        await Item.findByIdAndDelete(id);
         res.status(200)
             .json({ data: item });
     } catch (err) {
